+++ pybit/institutional_loan.py
@@ -8,6 +8,7 @@ class InstitutionalLoan(str, Enum):
     GET_REPAYMENT_ORDERS = "/v5/ins-loan/repaid-history"
     GET_LTV = "/v5/ins-loan/ltv-convert"
     BIND_OR_UNBIND_UID = "/v5/ins-loan/association-uid"
+    REPAY_LOAN = "/v5/ins-loan/repay-loan"
 
     def __str__(self) -> str:
         return self.value