+++ pybit/trade.py
@@ -13,6 +13,7 @@ class Trade(str, Enum):
     BATCH_CANCEL_ORDER = "/v5/order/cancel-batch"
     GET_BORROW_QUOTA = "/v5/order/spot-borrow-check"
     SET_DCP = "/v5/order/disconnected-cancel-all"
+    PRE_CHECK_ORDER = "/v5/order/pre-check"
 
     def __str__(self) -> str:
         return self.value